% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kijiji.R
\name{kijiji_scrape}
\alias{kijiji_scrape}
\title{Scrape rental unit data from Kijiji for Ottawa, ON}
\usage{
kijiji_scrape(num_pages = 10, old_urls = NA)
}
\arguments{
\item{num_pages}{How many pages to scrape? Default 10.}

\item{old_urls}{Optional. List of old URLs already scraped so we don't scrape
them again.}
}
\value{
A `tbl_df` with rental unit data.
}
\description{
Scrape rental unit data from Kijiji for Ottawa, ON
}
\examples{
\dontrun{apts <- kijiji_scrape(num_pages = 1)}
}
