% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padmapper.R
\name{padmapper_scrape}
\alias{padmapper_scrape}
\title{Scrape apartments from PadMapper within Ottawa}
\usage{
padmapper_scrape(
  old_data = NA,
  lat_min = 44.9,
  lat_max = 45.6,
  lng_min = -76.4,
  lng_max = -75.4,
  polite_pause = 5,
  verbose = TRUE
)
}
\arguments{
\item{old_data}{Optional tbl_df previous collected with this function. New
results will be added, minus duplicates.}

\item{lat_min}{Numeric. Minimum latitude for bounding box. Default 45.6.}

\item{lat_max}{Numeric. Maximum latitude for bounding box. Default 44.9.}

\item{lng_min}{Numeric. Minimum longitude for bounding box. Default -76.4.}

\item{lng_max}{Numeric. Maximum longitude for bounding box. Default -75.4.}

\item{polite_pause}{Numeric value in seconds to pause beween API calls. Default is 5.}

\item{verbose}{Boolean. Would you like lots of updates to the console? Default is TRUE.}
}
\value{
A tbl_df with apartment data scraped from PadMapper.
}
\description{
Scrape apartments from PadMapper within Ottawa
}
\examples{
\dontrun{
apartments <- padmapper_scrape()
}
}
