% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtorca.R
\name{realtor_scrape}
\alias{realtor_scrape}
\title{Scrape rental units from Realtor.ca}
\usage{
realtor_scrape(
  max_pages = 5,
  lat_max = 45.61892,
  lon_max = -75.02541,
  lat_min = 44.87931,
  lon_min = -76.57449
)
}
\arguments{
\item{max_pages}{Number of pages to capture.}

\item{lat_max}{Maximum latitude. Default values give Ottawa region.}

\item{lon_max}{Maximum longitude.}

\item{lat_min}{Minimum latitude.}

\item{lon_min}{Minimum longitude.}
}
\value{
A tbl_df containing unique rental units in the bounding box.
}
\description{
Scrape rental units from Realtor.ca
}
