% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rentalsca.R
\name{rentalsca_scrape}
\alias{rentalsca_scrape}
\title{Scrape rental unit information from Rentals.ca}
\usage{
rentalsca_scrape(
  cities = c("ottawa", "cumberland-on", "nepean", "gloucester", "kanata")
)
}
\arguments{
\item{cities}{Character vector. Must match rentals.ca's internal naming.
Defaults to c("ottawa", "cumberland-on", "nepean", "gloucester", "kanata")}
}
\value{
A `tbl_df` with rental unit information.
}
\description{
Scrape rental unit information from Rentals.ca
}
